Set up generation of HTML documentation from docstrings
The setup is a bit dodgy with a docs generator that hasn't been updated in
2 years, a heavily modified template that might break with any change to the
generator tool and no continuous integration, but I quite like the simplicity
and look of the resulting documentation. If it doesn't work out, I'll switch to
something more established later - for now it should be fine. After all this is
a small codebase.