Make wavenumber determine FWHM in filter_by_wavenumber
Previously the wavelength corresponding to the given wavenumber determined the
full width of the window. The change aligns the implementation with the
original intention of filtering like Ghinassi et al. (2018, 2020). Things
worked out before because the dominant wavenumber extracted from the meridional
wind is half that of the corresponding FALWA field so the filter widths were
sized as intended. Now the v-based wavenumber is doubled in the falwa_filtered
property of State before it is given to filter_by_wavenumber to compensate for
the change.

filter_by_wavenumber now also accepts zonally symmetric wavenumber fields as
1-dimensional nlat-sized arrays and is faster as it omits computation of
smoothed fields that are not required.